+++ apps/Event/groupAdmin/WM.js
@@ -1,5 +1,5 @@
-import { Check, Data, common, log, UCPr, Path, file } from '../../components/index.js'
-import { UCEvent } from '../../model/index.js'
+import { Check, Data, common, log, UCPr, Path, file } from '../../../components/index.js'
+import { UCEvent } from '../../../model/index.js'
 import { segment } from 'icqq'
 import _ from 'lodash'
 
@@ -25,20 +25,24 @@ class UCWelcome extends UCEvent {
     this.sub_type = 'increase'
   }
 
-  async accept(e, isGlobal) {
-    if (!this.isOpen || this.B) return false
-    if (this.user_id === this.qq) return log.white(`[新增群聊]${this.groupId}`)
+  async accept(e, isGlobal, isView = false) {
+    if (!isView) {
+      if (!this.isOpen || this.B) return false
+      if (this.user_id === this.qq) return log.white(`[新增群聊]${this.groupId}`)
+    }
     if (isGlobal) this.Cfg = UCPr.defaultCFG.GAconfig.welcome
     const replyMsg = []
+    if (isView && !this.isOpen) replyMsg.push(`注意：该群（${this.groupId}）未开启UC群管或入群欢迎\n`)
     if (this.Cfg.isAt) replyMsg.push(segment.at(this.userId))
     if (this.Cfg.isAvatar) {
       const avatarUrl = common.getAvatarUrl(this.userId)
       replyMsg.push(segment.image(avatarUrl))
     }
     const info = `<${e.nickname ?? this.userId}>（${this.userId}）`
+    log.white(`[群员增加]${info}`)
     const message = isGlobal ? globalWelcome : file.JSONreader(Path.get('WM', this.groupId, 'welcome.json')) ?? globalWelcome
     replyMsg.push(common.makeMsg(message, 'info', info))
-    return await common.sendMsgTo(this.groupId, replyMsg, 'Group')
+    return await common.sendMsgTo(isView ?? this.groupId, replyMsg, 'Group')
   }
 }
 
@@ -56,26 +60,31 @@ class UCMourn extends UCEvent {
     this.sub_type = 'decrease'
   }
 
-  async accept(e, isGlobal) {
-    if (!this.isOpen || this.B) return false
-    if (e.dismiss) return log.white(`[群聊解散]${this.groupId}`)
-    if (this.user_id === this.qq) {
-      if (e.operator_id !== this.qq) return log.white(`[被踢出群聊]${this.groupId}，操作者：${e.operator_id}`)
-      return log.white(`[退出群聊]${this.groupId}`)
+  async accept(e, isGlobal, isView = false) {
+    if (!isView) {
+      if (!this.isOpen || this.B) return false
+      if (e.dismiss) return log.white(`[群聊解散]${this.groupId}`)
+      if (this.user_id === this.qq) {
+        if (e.operator_id !== this.qq) return log.white(`[被踢出群聊]${this.groupId}，操作者：${e.operator_id}`)
+        return log.white(`[退出群聊]${this.groupId}`)
+      }
     }
+    if (isGlobal) this.Cfg = UCPr.defaultCFG.GAconfig.welcome
     const { card, nickname } = e.member ?? {}
     const info = `<${card || nickname || this.userId}>（${this.userId}）`
     const replyMsg = []
-    if (isGlobal) this.Cfg = UCPr.defaultCFG.GAconfig.welcome
+    if (isView && !this.isOpen) replyMsg.push(`注意：该群（${this.groupId}）未开启UC群管或退群通知\n`)
     if (this.Cfg.isAvatar) {
       const avatarUrl = common.getAvatarUrl(this.userId)
       replyMsg.push(segment.image(avatarUrl))
     }
     if (e.operator_id === this.userId) {
+      log.white(`[群员退群]${info}`)
       const message = isGlobal ? globalMourn : file.JSONreader(Path.get('WM', this.groupId, 'mourn.json')) ?? globalMourn
       const replyMsg = common.makeMsg(message, 'info', info)
-      return await common.sendMsgTo(this.groupId, replyMsg, 'Group')
+      return await common.sendMsgTo(isView ?? this.groupId, replyMsg, 'Group')
     } else if (e.operator_id !== this.qq) {
+      log.white(`[群员被踢]${info} 操作人：${e.operator_id}`)
       return await common.sendMsgTo(this.groupId, `${info}被管理员${e.operator_id ?? ''}踢出群聊`, 'Group')
     }
     log.white(`[机器人踢人]${info}`)
@@ -92,11 +101,11 @@ class UCWMset extends UCEvent {
       dsc: 'UC群管·进退群通知',
       rule: [
         {
-          reg: /^#?(UC)?查看(全局)?(入群欢迎|退群通知).*/i,
+          reg: /^#?(UC)?查看(全局)?(入群欢迎|退群通知)(\s*\d{5,10})?$/i,
           fnc: 'view'
         },
         {
-          reg: /^#?(UC)?修改(全局)?(入群欢迎|退群通知).*/i,
+          reg: /^#?(UC)?修改(全局)?(入群欢迎|退群通知)(\s*\d{5,10})?$/i,
           fnc: 'set'
         }
       ]
@@ -113,6 +122,11 @@ class UCWMset extends UCEvent {
       this.jsonPath = WM[this.type]
       this.data = file.JSONreader(this.jsonPath) ?? []
       return
+    }
+    const numMatch = this.msg.match(/\d+/)
+    if (!this.isGlobal && this.GM && numMatch) {
+      this.groupId = numMatch[0]
+      this.isAP = this.groupId
     }
     this.floderPath = Path.get('WM', this.groupId)
     this.jsonPath = Path.join(this.floderPath, `${this.type}.json`)
@@ -136,6 +150,8 @@ class UCWMset extends UCEvent {
   async view(e) {
     if (!this.isOpen) return e.reply(`请先开启UC群管及${this.string}`)
     if (!this.verifyPermission(this.Cfg.use)) return false
+    const oriGroup = e.group_id
+    e.group_id = this.groupId
     let app
     if (this.type === 'welcome') {
       app = new UCWelcome(e)
@@ -144,25 +160,28 @@ class UCWMset extends UCEvent {
       e.operator_id = this.userId
       app = new UCMourn(e)
     }
-    return await app.accept(e, this.isGlobal)
+    return await app.accept(e, this.isGlobal, oriGroup)
   }
 
   async set(e) {
     if (!this.isOpen) return e.reply(`请先开启UC群管及${this.string}`)
     if (!this.verifyPermission(this.Cfg.use)) return false
     if (this.isGlobal && !this.verifyLevel(4)) return false
-    const { type, string, floderPath, jsonPath, isGlobal, data } = this
-    e.oriData = { type, string, floderPath, jsonPath, isGlobal, data }
+    const { type, string, floderPath, jsonPath, isGlobal, data, isAP } = this
+    e.oriData = { type, string, floderPath, jsonPath, isGlobal, data, isAP }
     this.setFunction()
-    return this.reply(`请发送要修改的${isGlobal ? '全局' : ''}${this.string}\n注：info会被替换为“用户名（QQ）”`)
+    return this.reply(`请发送要修改的${isGlobal ? '全局' : (isAP ? `群${isAP}` : '本群')}${this.string}\n注：info会被替换为“用户名（QQ）”\n支持文字、图片、艾特组合`)
   }
 
   async setReply(e) {
     if (this.isCancel()) return
     if (!e.message) return this.finishReply('错误的内容')
     const { oriData } = this.getContext().setReply
-    if (_.isEmpty(oriData)) return this.errorReply()
-    const { type, floderPath, jsonPath, string, isGlobal, data } = oriData
+    if (_.isEmpty(oriData)) {
+      this.finish(this.setFnc)
+      return this.errorReply()
+    }
+    const { type, floderPath, jsonPath, string, isGlobal, data, isAP } = oriData
     const reply = []
     let imgCount = 0
     Check.floder(floderPath, true)
@@ -190,7 +209,7 @@ class UCWMset extends UCEvent {
     }
     if (_.isEmpty(reply)) return this.finishReply('无有效参数，请重新修改')
     file.JSONsaver(jsonPath, reply)
-    return this.finishReply(`修改${isGlobal ? '全局' : ''}${string}成功\n可通过#查看${string}\n查看效果`)
+    return this.finishReply(`修改${isGlobal ? '全局' : (isAP ? `群${isAP}` : '本群')}${string}成功\n可通过#查看${string}${isAP ?? ''}\n查看效果`)
   }
 }